Migrate tool_shed_registry to Galaxy.
